Refactor : Change Repository to JpaRepository
코드의 간소화를 위해 JpaRepository로 변경하였습니다.

이전 repository는 팀원들과의 논의 후 삭제할지 남길지를 결정할 예정입니다.
