+++ studycafe/src/main/java/com/scascanner/studycafe/login/repository/UserRepository.java
@@ -1,34 +1,10 @@
 package com.scascanner.studycafe.login.repository;
 
 import com.scascanner.studycafe.domain.User;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-import javax.persistence.EntityManager;
 import java.util.List;
 
-@Repository
-@RequiredArgsConstructor
-public class UserRepository {
-
-    private final EntityManager em;
-
-    public void save(User user){
-        em.persist(user);
-    }
-
-    public User findOne(Long id){
-        return em.find(User.class, id);
-    }
-
-    public List<User> findAll() {
-        return em.createQuery("select u from User u", User.class)
-                .getResultList();
-    }
-
-    public List<User> findByName(String name){
-        return em.createQuery("select u from User u where u.name = :name", User.class)
-                .setParameter("name", name)
-                .getResultList();
-    }
+public interface UserRepository extends JpaRepository<User, Long> {
+    List<User> findByName(String name);
 }
